backend/requirements.txt
fastapi>=0.115.0
uvicorn[standard]>=0.35.0
pydantic>=2.11.7
sqlalchemy[asyncio]>=2.0.30
asyncpg>=0.30.0
python-dotenv>=1.0.0
alembic>=1.15.0
pytest>=8.0.0
pytest-asyncio>=0.25.0
httpx>=0.28.0
aiosqlite>=0.20.0
